カスタムのresourceクラスを作れるようにした
