+++ README.md
@@ -217,6 +217,14 @@ tryDecoration()
 ```
 
 
+### Define Custom Resource class
+
+To define custom resource,  extends `ClayResource` class and use `.fromDriver()` method to create new instance
+
+```javascript
+
+```
+
 <!-- Section from "doc/guides/03.Advanced Usage.md.hbs" End -->
 
 <!-- Section from "doc/guides/10.API Guide.md.hbs" Start -->
@@ -226,7 +234,7 @@ tryDecoration()
 API Guide
 -----
 
-+ [clay-resource@3.0.1](./doc/api/api.md)
++ [clay-resource@3.0.2](./doc/api/api.md)
   + [create(args)](./doc/api/api.md#clay-resource-function-create)
   + [fromDriver(driver, nameString, options)](./doc/api/api.md#clay-resource-function-from-driver)
   + [ClayResource](./doc/api/api.md#clay-resource-class)
