+++ doc/guides/03.Advanced Usage.md.hbs
@@ -18,3 +18,12 @@ To add some custom logic before/after resource handling, use `.decorate(methodNa
 ```javascript
 {{{read 'example/example-decorate.js'}}}
 ```
+
+
+### Define Custom Resource class
+
+To define custom resource,  extends `ClayResource` class and use `.fromDriver()` method to create new instance
+
+```javascript
+{{{read 'example/example-class.js'}}}
+```
