+++ lib/clay_resource.js
@@ -25,6 +25,8 @@ const { decorate, create } = require('clay-entity')
 const co = require('co')
 const ResourceEvents = require('./resource_events')
 const clayResourceName = require('clay-resource-name')
+const { DriverSpec } = require('clay-constants')
+const { RESOURCE_BINDABLE_METHODS } = DriverSpec
 const {
   cloneMix,
   inboundMix,
@@ -561,6 +563,23 @@ class ClayResource extends ClayResourceBase {
       yield s.create(attributes)
       return yield s.of(attributes)
     })
+  }
+
+  /**
+   * @function fromDriver
+   * @param {Driver} driver - Driver to bind
+   * @param {string} nameString - Resource name string
+   * @param {Object} [options={}] - Optional settings
+   * @returns {ClayResource} Resource instance
+   */
+  static fromDriver (driver, nameString, options = {}) {
+    const Clazz = this
+    let resourceName = String(clayResourceName(nameString))
+    let bounds = RESOURCE_BINDABLE_METHODS.split(',')
+      .reduce((bounds, methodName) => Object.assign(bounds, {
+        [methodName]: driver[ methodName ].bind(driver, resourceName)
+      }), {})
+    return new Clazz(resourceName, bounds, options)
   }
 }
