+++ lib/from_driver.js
@@ -4,7 +4,7 @@
  * @param {Driver} driver - Driver to bind
  * @param {string} nameString - Resource name string
  * @param {Object} [options={}] - Optional settings
- * @returns {ClayResource} Created class
+ * @returns {ClayResource} Resource instance
  * @example
  * const { fromDriver } = require('clay-resource')
  * const { SqliteDriver } = require('clay-driver-sqlite')
@@ -15,19 +15,11 @@
  */
 'use strict'
 
-const { DriverSpec } = require('clay-constants')
-const { RESOURCE_BINDABLE_METHODS } = DriverSpec
-const clayResourceName = require('clay-resource-name')
-const create = require('./create')
+const ClayResource = require('./clay_resource')
 
 /** @lends fromDriver */
 function fromDriver (driver, nameString, options = {}) {
-  let resourceName = String(clayResourceName(nameString))
-  let bounds = RESOURCE_BINDABLE_METHODS.split(',')
-    .reduce((bounds, methodName) => Object.assign(bounds, {
-      [methodName]: driver[ methodName ].bind(driver, resourceName)
-    }), {})
-  return create(resourceName, bounds, options)
+  return ClayResource.fromDriver(driver, nameString, options)
 }
 
 module.exports = fromDriver
